+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Extended query result object in stat mode to provide a result value of the query execution as `value` property (i.e. `jora(query, { stat: true })().value`)
+- Renamed `SortingFunction` AST node type into `CompareFunction`
 - Added nullish coalescing operator (`??`)
 - Added `replace()` method
 - Added `min()` and `max()` methods